// frontend/oneCheck/src/model/originPreparationSheetDto.ts
/**
 * OneCheck.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DateOnly } from './dateOnly';
import { TimeOnly } from './timeOnly';


export interface OriginPreparationSheetDto { 
    checkSheetId?: string;
    date?: DateOnly;
    time?: TimeOnly;
    sealNumber?: string | null;
    signature?: string | null;
    name?: string | null;
    batteryStatusInPercent?: number;
    uldNumber?: string | null;
    awbNumber?: number;
    isBatteryCharged?: boolean;
    isContainerDamaged?: boolean;
    routing?: Array<string> | null;
    goodsPrecooledAtTemps?: number;
    uldPrecooledTemp?: number;
    tempSetting?: number;
    acceptedTempRange?: number;
    emergencyPhone?: string | null;
}